refactor: split shard stuff away from dir
